framework/requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
cliapp==1.0.9
constantly==15.1.0
cryptography==3.2
cssutils==1.0.2
daphne==2.5.0
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-extensions==3.0.9
django-gmailapi-backend==0.1.0
django-model-utils==4.0.0
djangorestframework==3.12.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.4
hiredis==1.1.0
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==0.6.2
oauthlib==3.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.11
zope.interface==5.1.2
